requirements.txt
Pillow>=10.0.0
PyGObject>=3.40.0
pycairo>=1.20.0 